fix: added error for Cayo 'name' prop
Right now it exits the program, but I'd rather it just log the issue.
However, this error never gets to the browser because the Cayo component
is part of the prerender, and accessing the logger from the Cayo
component would be a pain (not to mention it would get lost amongst the
other images pretty easily).

Still need to do a little more error checking with the component
modules themselves, to handle duplicate named files.

worked on #17