add template file CSS to output
Works the same as page CSS for the most part, except it's an additional
dependency of the page. If there is no CSS for the template, or
`internal: false` in the config, no CSS file will be written. If
`internal: true`, it'll be added in a style tag, just like page CSS.

fixes #30