Added better config handling for dev command
Also handled the user-defined viteConfig (which is defined within the
cayo config) a bit differently, in order to give precedence to the user
config over the internal vite config cayo uses. The merge uses a combine
merge function in order to handle the plugins array properly. This also
makes it so vite-plugin-svelte is always included, except should allow
users pass different options specifically into that plugin. Letting
users customize the svelte options was more important than keeping the
defaults, although `{ hydratable: true }` is required for the cayo
to work properly, so that needs to be documented as a requirement.

fixes #31