added temp fix for component updates not being reflected on pages
Non-Cayo components (regular, static render-only components) didn't any
pages using them to get re-prerendered, so their updates wouldn't show
up. In addition to that, their cached versions were sticking around from
there initial import on a new run of cayo dev. Essentially, you had to
restart the dev server in order to see changes to components reflected
on the pages using them.

Now, when a component is changed it triggers a rebuild of all pages and
their require cache is deleted in `pages.js`. In the future I'd want to
figure out a way to only rebuild the pages where the component is used.

This isn't as simple as I'd like it to be, because they're svelte files
and not JS files, so exisitng dependency tree tools probably won't work.