feat: cayo ids are generated again
But I changed some stuff. Cayo component src prop assumes it's relative
to the component dir. This makes things a lot easier, and you don't have
to write it as a true import, which is nice (albeit, maybe weird, but
either way the syntax is diff than actually importing a component)

Also should consider changing the prop from src to something else,
because rollup freaks out with this: `<Cayo src={"any string"} />`
even though that's valid Svelte. I think it sees src and tries to do
something else with it.