fix(#37): support preprocessors for cayos
In order to get the preprocessed CSS, if any, I had to move the cayo
component compilation. It used to be within the cli, like pages and
the template is compiled. In order to get the logic flow right, it
needed to be done in prerender, right after the cayo is detected during
post-process.

Still need to clean up old references to compile.cayos in cli files.